requirement.txt
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
einops==0.7.0
executing==2.0.1
fastapi==0.103.1
filelock==3.13.1
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.56.2
grpcio-tools==1.56.2
gym==0.10.9
h11==0.14.0
idna==3.4
ifcfg==0.24
imageio==2.32.0
imageio-ffmpeg==0.4.9
ipympl==0.9.3
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.0
jedi==0.19.1
Jinja2==3.1.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
numpy-quaternion==2022.4.3
oauthlib==3.2.2
openai==0.27.8
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.76
openni==2.3.0
packaging==23.2
paho-mqtt==1.6.1
parso==0.8.3
pexpect==4.8.0
Pillow==10.0.0
plotly==5.16.1
primesense==2.2.0.30.post5
prompt-toolkit==3.0.39
protobuf==4.23.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pyglet==2.0.10
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
torch==2.1.1
torch-scatter==2.1.2
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.65.0
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==2.0.4
uvicorn==0.23.2
wcwidth==0.2.9
Werkzeug==3.0.1
widgetsnbextension==4.0.9
yacs==0.1.8
yarl==1.9.2
